requirements.txt
Flask==3.0.0
openrouteservice==2.3.0
opencage==2.3.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1